+++ RELEASE NOTES.md
@@ -1,5 +1,12 @@
 RELEASE NOTES
 
+# 0.4.6
+
+* Secondary fields should be focused after opening them
+* Switch report fields after two (or three) digits
+* Better handling of downloaded data files (e.g. POTA database)
+* Offline POTA lookups
+
 # 0.4.5
 
 * Deleted QSOs are not exported as ADIF